+++ wp-content/themes/konocti/page_json.php
@@ -14,12 +14,10 @@
 
 namespace Ucoast\Konocti;
 
-//header( 'Content-type: application/json' );
+header( 'Content-type: application/json' );
 
 // Needed if you want to manually browse to this location for testing
 define('WP_USE_THEMES', false);
-
-
 
 // WP_Query arguments
 $args = array (
@@ -38,31 +36,37 @@ $loop = new \WP_Query( $args );
 //the loop
 while ( $loop->have_posts() ) : $loop->the_post();
 
-    if( get_post_type() == 'promotion') {
+    $type = get_post_type();
+
+    if( $type == 'promotion') {
         $group = get_field('promotion');
     }
 
-    else if( get_post_type() == 'event') {
+    else if( $type == 'event') {
         $group = get_field('event');
     }
 
     // create an array if there is more than one result
 
-    if($group['status'] == 'active') {
+    if( $group['status'] == 'active' ) {
 
-        $date_picker = $group['date'];
-        $day = substr($date_picker, -2, 2);
-        // if first character = 0, remove 0
-        $month = substr($date_picker, -4, 2);
+        $date_picker = get_field( 'date' );
+        $date_picker = substr( $date_picker, 0, 8 );
+        $day = ltrim( substr( $date_picker, -2, 2 ), '0' );
+        $month = ltrim( substr ($date_picker, -4, 2 ), '0' );
         $year = substr($date_picker, 0, 4);
+        $url = get_permalink();
 
         $raw_data[] = array(
-            'title' => $group['title'],
-            'day' => $day,
-            'month' => $month,
-            'year' => $year,
-            'start' => $group['start'],
-            'end' => $group['end']
+            'title'     => $group['title'],
+            'subtitle'  => $group['short_description'],
+            'day'       => $day,
+            'month'     => $month,
+            'year'      => $year,
+            'start'     => $group['start'],
+            'end'       => $group['end'],
+            'url'       => $url,
+            'event_type'      => $type
         );
     }
 
@@ -78,31 +82,36 @@ function format_events($events) {
     foreach ($events as $event) {
 
         if ( array_key_exists ( $event['year'], $data) ) {
-
             if ( array_key_exists( $event['month'], $data[$event['year']] ) ) {
+                // add to existing day array
                 if ( array_key_exists( $event['day'], $data[$event['year']][$event['month']] )) {
                     $data[$event['year']][$event['month']][$event['day']][] =
 
                             array(
                             'startTime' => $event['start'],
                             'endTime'   => $event['end'],
-                            'text'      => $event['title']
-
+                            'text'      => $event['title'],
+                            'url'       => $event['url'],
+                            'subtitle'  => $event['subtitle'],
+                            'eventType' => $event['event_type']
                         );
                 }
-
+                // create a new array for day
                 else {
                     $data[$event['year']][$event['month']][$event['day']] =
                         array (
                             array (
                                 'startTime' => $event['start'],
                                 'endTime'   => $event['end'],
-                                'text'      => $event['title']
+                                'text'      => $event['title'],
+                                'url'       => $event['url'],
+                                'subtitle' => $event['subtitle'],
+                                'eventType' => $event['event_type']
                             )
                         );
                 }
             }
-
+            // create a new array for month, day
             else {
                 $data[$event['year']][$event['month']] =
                     array (
@@ -111,13 +120,16 @@ function format_events($events) {
                             array (
                                 'startTime' => $event['start'],
                                 'endTime'   => $event['end'],
-                                'text'      => $event['title']
+                                'text'      => $event['title'],
+                                'url'       => $event['url'],
+                                'subtitle' => $event['subtitle'],
+                                'eventType' => $event['event_type']
                             )
                         )
                     );
             }
         }
-
+        // create a new array for year, month, day
         else {
             $data[$event['year']] = array( $event['month'] =>
                 array(
@@ -126,7 +138,10 @@ function format_events($events) {
                         array (
                             'startTime' => $event['start'],
                             'endTime'   => $event['end'],
-                            'text'      => $event['title']
+                            'text'      => $event['title'],
+                            'url'       => $event['url'],
+                            'subtitle' => $event['subtitle'],
+                            'eventType' => $event['event_type']
                         )
                     )
                 )
@@ -134,75 +149,8 @@ function format_events($events) {
         }
     }
 
-    // sort by year key
-
     return $data;
 }
-/*
-function format_events($events) {
-
-    $data[] = array();
-
-    foreach ($events as $event) {
-
-        foreach ($data as $year) {
-            if ( array_key_exists ( $year, $data ) ) {
-                $event_year = $data[$year];
-
-                foreach ( $event_year as $month ) {
-                    if( array_key_exists( $month, $data[$year] ) ) {
-                       $event_month = $data[$year][$month];
-
-                       foreach ( $event_month as $day ) {
-                           if( array_key_exists( $day, $data[$year][$month] ) ) {
-                               $data[$year][$month][$day] = array(
-                                   'startTime' => $event['start_time'],
-                                   'endTime'   => $event['end_time'],
-                                   'text'      => $event['title']
-                               );
-                           }
-
-                           else {
-                               $data[$year][$event['month']][$event['day']] = array(
-                                   'startTime' => $event['start_time'],
-                                   'endTime'   => $event['end_time'],
-                                   'text'      => $event['title']
-                               );
-                           }
-                       }
-                    }
-
-                    else {
-                        $data[$year] = $event['month'];
-                        $data[$year][$event['month']] =  $event['day'];
-                        $data[$year][$event['month']][$event['day']] = array(
-                            'startTime' => $event['start_time'],
-                            'endTime'   => $event['end_time'],
-                            'text'      => $event['title']
-                        );
-                    }
-                }
-            }
-
-            else {
-                $data[] = $year;
-                $data[$year] = $event['month'];
-                $data[$year][$event['month']] =  $event['day'];
-                $data[$year][$event['month']][$event['day']] = array(
-                    'startTime' => $event['start_time'],
-                    'endTime'   => $event['end_time'],
-                    'text'      => $event['title']
-                );
-            }
-        }
-    }
-
-    return $data;
-}
-*/
-//d( format_events( $raw_data, $data ) );
-
-
 
 // output
 echo json_encode( format_events( $raw_data ) );